Add is:powerfulreward search and improve pursuit searching
