+++ src/app/progress/Pursuit.tsx
@@ -4,12 +4,43 @@ import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import ItemExpiration from 'app/item-popup/ItemExpiration';
 import PursuitItem from './PursuitItem';
 import { percent } from 'app/shell/filters';
+import ItemObjectives from 'app/item-popup/ItemObjectives';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions.service';
+import { RootState } from 'app/store/reducers';
+import { searchFilterSelector } from 'app/search/search-filters';
+import { connect } from 'react-redux';
+import classNames from 'classnames';
+
+// Props provided from parents
+interface ProvidedProps {
+  item: DimItem;
+  defs: D2ManifestDefinitions;
+}
+
+// Props from Redux via mapStateToProps
+interface StoreProps {
+  isNew: boolean;
+  searchHidden?: boolean;
+}
+
+function mapStateToProps(state: RootState, props: ProvidedProps): StoreProps {
+  const { item } = props;
+
+  const settings = state.settings;
+
+  return {
+    isNew: settings.showNewItems ? state.inventory.newItems.has(item.id) : false,
+    searchHidden: !searchFilterSelector(state)(item)
+  };
+}
+
+type Props = ProvidedProps & StoreProps;
 
 /**
  * A Pursuit is an inventory item that represents a bounty or quest. This displays
  * a pursuit tile for the Progress page.
  */
-export default function Pursuit({ item }: { item: DimItem }) {
+function Pursuit({ item, defs, isNew, searchHidden }: Props) {
   const expired = showPursuitAsExpired(item);
 
   const nonIntegerObjectives = item.objectives
@@ -23,10 +54,13 @@ export default function Pursuit({ item }: { item: DimItem }) {
     (nonIntegerObjectives.length === 1 && !nonIntegerObjectives[0].boolean);
 
   return (
-    <div className="milestone-quest" key={item.index}>
+    <div
+      className={classNames('milestone-quest', { 'search-hidden': searchHidden })}
+      key={item.index}
+    >
       <div className="milestone-icon">
         <ItemPopupTrigger item={item}>
-          <PursuitItem item={item} />
+          <PursuitItem item={item} isNew={isNew} />
         </ItemPopupTrigger>
         {!item.complete && !expired && showObjectiveProgress && (
           <span>
@@ -52,6 +86,8 @@ export default function Pursuit({ item }: { item: DimItem }) {
     </div>
   );
 }
+
+export default connect<StoreProps>(mapStateToProps)(Pursuit);
 
 /**
  * Should this item be displayed as expired (no longer completable)?
