+++ src/app/progress/PursuitItem.tsx
@@ -9,36 +9,9 @@ import pursuitComplete from 'images/pursuitComplete.svg';
 import pursuitExpired from 'images/pursuitExpired.svg';
 import trackedIcon from 'images/trackedIcon.svg';
 import { showPursuitAsExpired } from './Pursuit';
-import { RootState } from 'app/store/reducers';
-import { searchFilterSelector } from 'app/search/search-filters';
-import { connect } from 'react-redux';
 import { count } from 'app/util';
 
-// Props provided from parents
-interface ProvidedProps {
-  item: DimItem;
-}
-
-// Props from Redux via mapStateToProps
-interface StoreProps {
-  isNew: boolean;
-  searchHidden?: boolean;
-}
-
-function mapStateToProps(state: RootState, props: ProvidedProps): StoreProps {
-  const { item } = props;
-
-  const settings = state.settings;
-
-  return {
-    isNew: settings.showNewItems ? state.inventory.newItems.has(item.id) : false,
-    searchHidden: !searchFilterSelector(state)(item)
-  };
-}
-
-type Props = ProvidedProps & StoreProps;
-
-function PursuitItem({ item, isNew, searchHidden }: Props) {
+export default function PursuitItem({ item, isNew }: { item: DimItem; isNew: boolean }) {
   const isCapped = item.amount === item.maxStackSize && item.uniqueStack;
   const expired = showPursuitAsExpired(item);
   const showProgressBar =
@@ -50,7 +23,6 @@ function PursuitItem({ item, isNew, searchHidden }: Props) {
     (item.objectives.some((o) => o.displayStyle !== 'integer' && !o.boolean) ||
       count(item.objectives, (o) => o.boolean) > 1);
   const itemImageStyles = {
-    'search-hidden': searchHidden,
     [styles.tracked]: item.tracked
   };
   return (
@@ -77,5 +49,3 @@ function PursuitItem({ item, isNew, searchHidden }: Props) {
     </div>
   );
 }
-
-export default connect<StoreProps>(mapStateToProps)(PursuitItem);
